worked on the pitcher class,understood the remove range, it was different from what I initially thought even though the output was the same.

+++ Lemonade Stand/UserInterface.cs
@@ -46,10 +46,10 @@ namespace LemonadeStand
             return amountofIceCubes;
         }
 
-        public static double PricePerCup()
+        public static int PricePerCup()
         {
             Console.WriteLine(" Please enter your price per cup.");
-            double pricePerCup = Convert.ToInt32(Console.ReadLine());
+            int pricePerCup = Convert.ToInt32(Console.ReadLine());
             return pricePerCup;
         }
     }
